+++ Changelog.md
@@ -3,6 +3,7 @@
 ## NEXT
 
 * Fixed path issue in webpack config for certain JS libraries
+* Added option to export books from library
 
 ## 2.13.0